fix: multicall2 arb1 address, Rinkeby and Arbitrum Rinkeby default RPC endpoints
and refactoring the getProvider
