+++ src/constants.ts
@@ -128,9 +128,9 @@ const MULTICALL_ADDRESS: { [chainId in ChainId]: string } = {
 const MULTICALL2_ADDRESS: { [chainId in ChainId]: string } = {
   [ChainId.MAINNET]: '0x5ba1e12693dc8f9c48aad8770482f4739beed696',
   [ChainId.RINKEBY]: '0x5ba1e12693dc8f9c48aad8770482f4739beed696',
-  [ChainId.ARBITRUM_ONE]: '0xF718F2bd590E5621e53f7b89398e52f7Acced8ca',
+  [ChainId.ARBITRUM_ONE]: '0x80c7dd17b01855a6d2347444a0fcc36136a314de',
   [ChainId.XDAI]: '0xFAa296891cA6CECAF2D86eF5F7590316d0A17dA0',
-  [ChainId.ARBITRUM_RINKEBY]: '0x5ba1e12693dc8f9c48aad8770482f4739beed696',
+  [ChainId.ARBITRUM_RINKEBY]: '0x309e61A4c36a4a9f131f8844eA521F6384B6C9E3',
 }
 
 export {
