+++ src/entities/trades/uniswap-v2/contracts/index.ts
@@ -8,7 +8,7 @@ import { Pair } from '../../../pair'
 import { ChainId, MULTICALL2_ABI, MULTICALL2_ADDRESS } from '../../../../constants'
 import { UniswapV2RoutablePlatform } from '../../routable-platform/uniswap-v2-routable-platform'
 import { Interface } from '@ethersproject/abi'
-import { getProvider } from '../../curve/contracts/utils'
+import { getProvider } from '../../utils'
 import { Token } from '../../../../entities/token'
 import { TokenAmount } from '../../../../entities/fractions'
 import { BASES_TO_CHECK_TRADES_AGAINST } from '../constants'
